changed edi_new_object to edi and edi to edi_old
